Changed substX... to Replacable typeclass
There were a lot of functions which were doing the same stuff but on different
data types, that is, the propogation of the results from a unification.  

Created a new typeclass "Repclable" and provided instances for these types.

Leads to more concise code which more effectively communicates the 
"sameness" of this operation.  Maybe?