Allows for semantics in trees + Improves XMG lex selection
This lets us get rid of a tree anchoring hack in which we would take the
semantics from the lexical item and ignore the semantics of anchored tree
(SelectTAG).  It also removes a notational hack in which we named the families
xLEX_IDxFAMILY_NAME in order to point each tree back to the lexical item from
which it was selected.  We no longer retain this relationship.

Additional improvements
=======================

Geni
----
- Factorises lexical selection so that only the anchoring process is different
  (Merges quite a bit of redundant lex selection code together)
- Moved lexical item selection before the tree anchoring stuff
- Inserts all lexemes into the lexically selected tree name (as opposed to just
  the anchor).

GeniParsers
-----------
- Factorises sortSem so we don't get anymore accidents with semantic comparison
- Allows for raw identifiers as lexeme nodes in trees (not just string literals)