Fixed unix process issues with SelectTAG! (i hope).
Yay! I'd been struggling with this problem for half a year now!

The basic problem is that you have to read from the stderr of the subprocess,
or else it will block forever, waiting for something?  I guess what's
happening is that it's trying to write to a file handle that is only made
available when the parent process does hGetContents on it.