Corrected alpha-conversion bug
Variables in tree semantics were not being alpha-converted
this didn't really matter up till now because we had always 
assumed the semantics would be instantiated by the time we
get to the builder, but now that we are more liberal in the
kind of semantics we read, alpha-conversion needs to be
extended.

Refactored alpha conversion to use the replace function - this
is probably a whole lot less efficient, but it's likely to be 
a good bit less error prone.