(cky) fixed dispatch mechanism
Information was getting lost in the case where the dispatch filter
does not succeed.  For example, the redundancy filter fails if the
item is unique, but when it does, it should try to set the item id
of the new item.  Up to now, this new id was getting lost.

The fix consists of returning (Just newItem) on failure.  This way
we can pass the modified item on the next dispatch filter.  What's
handy about this is that we can also use the same mechanism for
top/bottom filtering.  The tbFailure filter fails if tb unification
succeeds, so it returns Just itemWithTbUnified.