(cky) Fix adjunction rule.
Do *not* copy the unified bottom into the node.  Keep in mind that this node
represents the upper half of the adjunction.  The bottom half is represented
by the auxiliary tree and its foot note.

NB: note that while we do calculate this bottom node to make sure unification
does not clash and to get a new list of substitutions, we do not actually use
it anywhere.