(cky) Calculate the automaton *after* derivation.
It doesn't work if you calculate it simultaneously because items may be merged
into after you've used them to calculate your automaton.  Besides, it makes the
code cleaner and doesn't concretely make a difference because Haskell is lazy
anyway.
  
Note: not yet used in the GUI.  I still have to get rid of the old code.

(Also: more rearranging and literate commenting of the code)