Force the mode on 'promettre' example trees.
I'm just trying to prevent overgeneration.  There are trees which explicitly
constrain for sentences with mode:inf, but these constraints are ignored if
the mode is unset.  So I give all sentences with an unset mode the default mode
FIXME, so that a real linguist can come clean them up.