Simplify polarity code.  Refactor calls to polarity stuff.
It's rather difficulte to seperate all this stuff, because my
simplifications to how the polarity code works has repercussions on how
you call it, and since in any case I'm changing the way you call it,
I might as well refactor the whole deal.

Before, the polarity code tried to operate on something I called a
TagLite.  I was under the mistaken impression that you needed to reduce
the data structures you were passing around to save memory or something,
but with lazy evaluation and this whole business of thunks and pointers,
it is actually not neccesary.  It just makes the code complicated for
nothing.  So this patch removes this process of converting TagElem to
TagLite and back.

One important change: before, we would do tricks to the TagElem whilst
converting back from TagLite.  Namely, if the TagLite's semantics refers
to one that we added during pronoun detection, we would strip this
semantics off and set the TagElem's index.  Now that we don't have a
conversion step, you get TagElems with a semantics that is not in the
input.  Argh.  So we return the new input semantics, including the
pronouns we inserted and everything is happy again.  The basic point
being that the polarity optimisation can introduce semantics.  Yuck.

The rest of it is just a refactor between CKY and Simple.


darcs-patches: H4sIAAAAAAAAA3NVKEotyElMTlXQ0y8uStZPT83L1C/OzC3ISdUPBlPupZl6ORnFCtGOulGJulXxBrqWsQrFqSWlBQH5OYlFmSWZqcUKBUWpnnmZJQBMBaRtTgAAAA==