(simple) Tidy and rearrange everything.
Attempt to organise code in a more readable, literate manner.
Strip trailing newlines.

I hope patches like this aren't in horribly bad taste.  This patch
is just meant to tidy everything up before I create the SimpleItem
data type to be used as a chart item in lieu of TagElem.