(cky) Scaled back overzealous bug fix on kids to parents rule.
I had thrown in a small semantic check to make sure that the semantics
of the kid items do not overlap.  This was to avoid overgeneration where
you substitute the same stuff in different places.  But of course, this
semantic check means you can *never* use the kids to parents rule,
because everything in the same tree forcibly overlaps with its own
semantics!  So we scale back the semantic overlap check to only enforce
non-overlapping on semantics other than the tree semantics.