(cky) Add an optional inference rule for ordered substitution.
I call this the nextLeaf rule.  Note that when this is enabled, we
also reduce the number of items with which we initialise the agenda.

Note: this changes the configuration settings.  "O" is no longer so
short for "ordered adjunction"; "Oa" is.  This is so that we can
introduce "Os" for "ordered substitution".