Remove substnodes and adjnodes in the definition of TagElem.
Move them to the definition of simple chart item.  These are things that are
computed (and used only for the chart and polarity stuff), whereas in TagElem,
it is preferable to restrict things that are actually defined in a grammar or
via lexical selection.

Another hope is that this would make haskell-outputted grammars compile faster.