(perf) Remove needles sort.nub on adjlist in subst rule (see comment!)
Interestingly, just removing the sort.nub hurts performance, which is weird
because the nodes are unique and order is meaningless.  So why does sorting
improve performance?  I think it's because it puts the core nodes first (the
tree receiving substitution), and core nodes are "harder" in the sense that
top/bot unification fails more easily on them.  So to get the same effect,
we simply change the order; adj2 before adj1.

Weird, huh?