Large overhaul of simple (see comment, perf)
When building derived trees, don't use GNode for the tree nodes; just use a
node name instead.  All the information we really need from a GNode is
redundant with the TagSite stuff anyway.  The only real use for it is in the
Gui. To make up for it, we include some extra information to let you map your
way back when you're in gui mode.

This leads to only a slight performance increase.  We allocate 5% less memory
(about 25M) on the sentence
 une femme gentille discute avec un ingénieur gentil excellent affectueux