Fix bug in fixPronoun?
During the semantic index counting mechanism, cap the number of indices we
insert into the tree semantics so that it never exceeds the number of extra
columns inserted.  This is only important now that we're representing the
semantics with a bit vector.  We want to avoid things sneaking into tree
semantics that aren't in the input semantics.