Faster replacement scheme (Thanks to Sebastien Hinderer).
Instead of doing a bunch of O(n) test and compares, we do a single
O(n^2) compilation of the replacement lists into a map and then
a bunch of cheaper O(log n) lookups.  There might even be a way to
reduce the lookup time, for example, by using collect to shrink the
map down to only the open variables in the data type.

Note that this currently uses more memory, though it seems to be
a bit faster.  But the code is cleaner in that I took out some
premature optimisations (which might also help explain the
increased memory).  I'm hoping now that by directly building
a Map during unification instead of a list, I can save some of
this memory.