Add a new field to testcases, tcOvergens.
This helps us keep track of sentences which are very clearly NOT what we
want to produce.  The idea is that the test suite writer might want
highlight certain red flags, i.e. "if your grammar produces this
sentence for this output, you are in big trouble".  Not used for
anything yet, but I might slip it into some test suite hackery that
we're using for ACL 2007 validation.