Fix bug in lexical selection filter.
The situation is this: sometimes the anchoring stage will return a tree
that has more literals in its semantics than its anchor.  Sometimes the 
"bonus" literals are rubbish, so to prevent them from entering the
lexical selection, we filter out any tree with a literal that isn't in
the input semantics.

This is fine in principle, but my implementation was wrong because I
was checking to see if it literally was in the input semantics.  If
there are stray variables floating around (i.e. that haven't been
replaced with a constant), you'll get spurious inequalities.  The
filter will think it's a rubbish literal, when in fact it really is
part of the semantics.

The correct procedure is thus not to check for equality but for
subsumption.  Bonus literals are accepted if they subsume the input
semantics -- that is if they unify with some literal in the semantics --
not just if they are a subset of it.