Be less clever about parsing files (but more robust).
The old behaviour was that if the file had an error, we would try
parsing it as a binary... which is stupid because sometimes it would
go into non-termination trying to binary parse a text file.  Also,
I did not make any attempt at recovering the error if it was indeed
a problem with the text file.

The new behaviour is to only parse as binary if the filenname has
extension genib


darcs-patches: H4sIAAAAAAAAA3NVKEotyElMTlXQ08/JTEpPzcvU9/MJ0HdPzfMEEZl6ORnFCtGOulGJulXxBrqWsQoFiUXFqW5F+blumTmp/kUhRZVOmXmJRZUICd/EyqRUiKB/kVtiZg4AbdRwUGYAAAA=