Fix morph lexicon string literal stuff.
Parsing for geniWord first consumes the '"' token.
Switching it around so that stringLiteral comes first should
make it so that stringLiterals with spaces in them can be
succesfully parsed.  On the other hand, we lose the ability
to parse tokens with only one quote in them, which may come
up in a bug report one day.