Deeper fix for unification bug.
My first fix was probably wrong (after a night sleeping on it).  The reason my
explanation did not make sense is that when you do top and bottom
unification, you should not be able to produce anything that is redundant with
the previous unification steps.  The _real_ bug was that the foot top node did
not receive all of the variable substitutions...

But then getting rid of this revealed something else, which is that if you
unify a variable with itself, you get a loop in the substitutions lists, which
makes it possible to put yourself into prependSubst error.

We really ought to rethink how we do these unifications so that we run into
fewer errors.