Bug fix: polarity filtering compensator was only looking for categories.
The compensator is the mechanism we use to make an exception for the S category
during polarity filtering.  Normally, you want your automata to start at 0 and
end at 0.  One crucial exception is the S category (all derived trees have to
have *a* root node, so you want your polarities to balance out as +cat:S).
I forgot to generalise this mechanism to work for non-category attributes.