Fix regression in polarity code and make it less error-prone.
The problem is that we model GConst as [String], and distinguish between the
case where there is no disjunction ([foo]) and where there is ([foo,bar,baz]).
But because this stuff is not immediately obvious, it's very easily to
inadvertenly concatenate some (non)-disjunctive stuff and treat things as
disjunctions that really ought to be constants.

The solution is use a newtype PolarityValue to prevent this errornous
concatenation from taking place.