Fix #31 - unification bug cause by failure to propagate.
We do top and bottom unification after adjunction; however, the
nodes that we were doing unification on did not have the latest
substitutions from the adjunction test applied to them.

I'm still not sure why this is a regression from GenI 0.17, but
that's a future question to research.

Note also that this bug would not have been caught without the
pattern-matching style allowed by languages such as Haskell or one of
the MLs.  We simply would have silently accepted the anomalous situation
and the user would have been none the wiser.  I consider this a
succesful application of the "Dead Programs Don't Lie" principle.
So yay Haskell!