Fix subtle GUI bug causing display of wrong warnings (or more serious).
The problem is in the use of imperative-style programming: basically I
was doing something which modified the program state (pstRef), but
passing in a stale pst to a function.

Perhaps the right thing to do is to go *more* imperative: always use
pstRefs whenever you have an IO function; don't try to rely on a fixed
pst?

Or alternatively, work harder in sequestering the IO from the pure stuff.
Hmmph...