#33 - Make the root feature flag required, with no defaults.
To be clear, I don't think the root feature is necessary unless you use
polarity filtering, but it can helpful performance-wise during chart generation
because of the root filter optimisation (now default).

A reason I want to make it required is that I don't users to be misled into
confusing performance gains due to the root filter optimisation, with those
that are due to polarity filtering.