Core support for variables with constraints on their possible values!
This is a pretty important change to the representation of GeniVal.
The old GeniVal looked something like this

  type Constant = String
  type Label    = String

  data GeniVal = GConst [Constant]
               | GVar Label
               | GAnon

The new GeniVal effectively changes the case of GVar to
                 GVar Label (Maybe [Constant])

In other words, this allows us to optionally associated the
variables with a disjunction of possible constants the variable
is allowed to unify with.

Note that the actual implementation is a looks rather different, but
I think that this difference is essentially superficial:

  data GeniVal = GeniVal (Maybe Label) (Maybe [Constant])

The idea is that all GeniVal can optionally have a label and a list
of constraints.  This makes constant values as equivalent to
anonymous variables plus constraints.

Note that this does not actually enable you to write GenI grammars with
such variables.  That is for a separate patch.