Basically get rid of null builder, moving all polarity stats to common.
I'm not sure if this has negative consequences for performance by forcing
stuff to be evaluated, though; but it seems to be less broken than the
previous state.  I'm not entirely clear on what the problem is, perhaps
some kind of difference between the input that both init functions consume.

The Builder init gets inputs which are pre-processed by preInit, and I'm
no longer sure how that affects things.