Slightly safer variant of feature structures.
This is a bit experimental, but it always seemed a bit silly to
check for uniqueness of features at runtime.  Also, it seems
like actually a good thing to avoid implying an inherent
ordering of the features, as a list would.  But is it worth the more
complicated type, and the loss of list comprehension syntax?
Our unification work relies on an imposed ordering too.