Simplify lexical selection (remove key-indexing)
Key indexing (from Carlos) seems like a good idea for performance,
but when we introduce lexical items with multi-literal semantics
and variables in the predicates, it becomes a bit of a liability.
For now, I prefer to just Keep It Simple.

Note that this should also fix a bug where lexical selection
returns the wrong result if the lexical item has only literals
with variable predicates in their semantics.

This patch also nukes some code along the way that's only used by
key indexing.