Slightly improve 'alphaConvert' when no ac is requested.
Our 'alpha conversion' consists in appending a suffix to all variable names
on the assumption that you use the same convention on all unifiable objects
whenever running it.  If you don't supply a suffix, alpha conversion itself
should be the same as id (a QuickCheck for this would be good)

Why would anybody call alphaConvert with a null suffix?  Because the name is
just a bit stale; it does more than alpha conversion (singleton anonymisation
for now).  May need to change this in a later patch