Simplify away use of multiparameter type classes in GV code.
This is related to recent work to eliminate the emptyFoo functions from
GenI. I discovered that we *were* using emptyTE, to convert SimpleItem
to TagElem so we could generate a graphviz visualisation for it.  The
basic change was to try generalising the GraphvizShow class to take
anything in the TagItem typeclass (extending it to know about trees),
but that got me into ambiguous instances mess.

The new GraphvizShow does not have a separate flag paramater (sorry) but
expects you to associate the item itself with the flag.  It's a bit
yuckier conceptually, but combined with some other changes to use custom
intermediary data structures (over complex anonymous tuples) results in
simpler code.

What a yak-shave!  All I wanted to do was implement GeniShow ILexEntry,
but doing that made me realise it would be good to close a hole in
ILexEntry by changing ilexeme to Data.FullList, and doing that made me
realise I should nuke emptyILE... and the rest is history.