Remove NLP.GenI.Btypes.
I'm still of two minds about this module.  I think the simplified imports
would be nice (albeit as NLP.GenI), by I find it hard to keep up with
remembering if something is exposed in the unified imports, much less to
say deciding if it should go in it.  So it just winds up being a horrible,
well-intentioned, inconsistent mess.