Switch back from SmallCheck to QuickCheck.
I ran into problems building the test suite with GHC 7.4.1.  The main issue is
that for the Serial Text instance I'd dumbly created, I apparantly now need a
Generic instance for Text, but that this cannot be auto-derived because the
Text constructor is not exposed.

I just figured I'd rather get the release out the door than deal with this
at the moment.