Don't rely on ranking flag to determine if we have ranking constraints. #changelog
This should only affect users of the library.
The batch/gui would not recognise that you had ranking constraints if
they were passed in directly from the library instead of via the flags.

This is ultimately pointing to some sort of code smell we need to address,
redudancy between the flags and the config state.